finally find error
