+++ BEMscripts/utilities/convert_mp4_to_mp3/start.js
@@ -1,0 +1,6 @@
+import convert from "./convert.js";
+
+const fileName = "./video.mp4";
+const saveToFileName = "audio.mp3";
+
+convert(fileName, saveToFileName);
